move fill_from_rcmip to new module
